% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Attributes}
\alias{Attributes}
\title{Attributes of OTUs}
\format{
A matrix with 3 columns and 8 rows.
}
\usage{
data(Attributes)
}
\description{
A matrix with 3 columns and 8 rows.
}
\details{
The first column contains OTU names:

"RBL636","rkd D5891 bcc","ylnC-48","iglK f12","BAeR G11","nuhj-25","HwrK-41","QecF*22"

The second column contains parents of the corresponding OTUs:

NA ,"RBL636","RBL636","ylnC-48","rkd D5891 bc","iglK f12","rkd D5891 bc","nuhj-25"

and the third column contains colors of corresponding OTUs.
}
\keyword{datasets}
