% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PopulationDataList}
\alias{PopulationDataList}
\title{Population/Abundance data of OTUs over time for "list" method}
\format{
A data frame with 3 columns and 382 rows.
}
\usage{
data(PopulationDataList)
}
\description{
This data is a small artificial data which contains information about population/abundance of
some artificial OTUs. When data.method in Muller.plot function is "list" then population.data
must be a matrix or data frame with 3 columns.
}
\details{
Here the first column contains OTU names:
"RBL636","rkd D5891 bcc","ylnC-48","iglK f12","BAeR G11","nuhj-25","HwrK-41","QecF*22"

The second column contains time steps in which the corresponding OTUs are present in the dynamics.

The third column contains populations/abundances of corresponding OTUs at corresponding times.
}
\keyword{datasets}
