% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PopulationDataTable}
\alias{PopulationDataTable}
\title{Population/Abundance data of OTUs over time for "table" method}
\format{
A matrix with 101 columns and 8 rows.
}
\usage{
data(PopulationDataTable)
}
\description{
This data is a small artificial data which contains information about population/abundance of
some artificial OTUs. When data.method in Muller.plot function is "table" then population.data
must be a matrix or data frame with N rows and T columns, which N is the number of OTUs and T
is the number of time steps or generations with provided abundances (or the number of samples). Here
N=8 and T=101.
}
\details{
"rownames" of this matrix/data frame must be OTU names:

"RBL636","rkd D5891 bcc","ylnC-48","iglK f12","BAeR G11","nuhj-25","HwrK-41","QecF*22"

and "colnames" must be time steps or generations.

The Matrix contains abundance of each OTU at each time step.
}
\keyword{datasets}
